path: Don't unwatch specs until inotify removes them
Otherwise, path units are inherently racy.  We might miss an inotify
event during the time that we're unwatching and reestablishing the watch
on a particular file.

For example, a simple path unit with PathExists=/tmp/foo will receive an
IN_ATTRIB event when the file is unlinked (since the inode link count is
being modified) and an IN_DELETE_SELF event when the inode entry is
actually removed.  If IN_ATTRIB is received first and we unwatch
/tmp/foo, we might miss IN_DELETE_SELF.  (This is its own problem that
the next commit addresses.)

It should be considered safe to unwatch a particular file once inotify
tells us that the watch was removed.  At that point we can safely no
longer expect to receive events from it.