path: Don't recheck PathExists on IN_ATTRIB events
PathExists files should only be rechecked under two conditions, both of
which indicate an actual change in the existence of an inode entry for a
given path on the filesystem:

(1) when we receive the initial watch event for a path

If a path does not previously exist, path_spec_fd_event() won't be able
to match the event watch descriptor (wd) to the PathSpec primary wd
(since primary_wd hasn't been set yet).  This typically means the event
is being triggered on the path's parent because the path is being
created.  inotify_mask is empty in this scenario.

(2) when we receive IN_DELETE_SELF or IN_MOVE_SELF events

When unlinking a file being watched by PathExists, we may receive an
IN_ATTRIB event before IN_DELETE_SELF.  If we recheck on the IN_ATTRIB,
it's likely that the inode entry for the file still actually exists, so
calling access() in path_spec_check_good() during recheck gives
misleading results.  This can cause the path's service to be erronously
retriggered when the path is actually being removed.